Add files via upload
In this project wew taken Data of the Cement to predict the Strength of the Cement on the basis of the previous one. In this data there are many parameter which are affect the strength of the cement. First we find the best parameter which are affecting more to the Strenght of the Cement. There are the parameter which are affecting  are Cement, Blast Furnace Slag, Fly Ash, Water, Superplasticizer, Coarse Aggregate, Fine Aggregate, Age. In this for the best cement strength all these feature have a specific values for that strength of Cement is strogest at that Values. In this we do many preprocess like remove duplicate value, outlier and many more other processing data.
We made Pipeline by which we give some models in pipeline and calculate their r2 score and on the basis of this r2 score we select a model out of these which is GradientBoostingRegressor and check its best parameter by using the gridsearchCV and then model is ready to make prediction which r2 score is nearly 0.92 which is very good.